Solve part 1 of day 9
